requirements.txt
#this file is made using cmd pip freeze > requirements.txt
# using pip install -r requirements.txt , all the packages will be installed on your device

alchemy==20.5
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
binary==1.0.0
bitarray==2.8.2
certifi==2023.7.22
chardet==3.0.4
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
docutils==0.20.1
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.0.12
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.8
passlib==1.7.4
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-daemon==2.2.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.22.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.25.11
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
